Split immigration.rs (610 lines) into immigration/ subdirectory
Split crates/simulation/src/immigration.rs into focused modules:
- types.rs: CityAttractiveness, ImmigrationStats resources and constants
- attractiveness.rs: compute_attractiveness system
- waves.rs: immigration_wave, spawn/emigration logic
- random.rs: tick_pseudo_random helper
- plugin.rs: ImmigrationPlugin registration
- tests.rs: all unit tests

All public items re-exported from mod.rs — no external import changes needed.
